Refactoring: Move KnownRuntimePack to a new type
