+++ src/Tasks/Microsoft.NET.Build.Tasks/ProcessFrameworkReferences.cs
@@ -704,46 +704,5 @@ namespace Microsoft.NET.Build.Tasks
                 return ShortFolderName;
             }
         }
-
-        internal struct KnownRuntimePack
-        {
-            ITaskItem _item;
-
-            public KnownRuntimePack(ITaskItem item)
-            {
-                _item = item;
-                TargetFramework = NuGetFramework.Parse(item.GetMetadata("TargetFramework"));
-                string runtimePackLabels = item.GetMetadata(MetadataKeys.RuntimePackLabels);
-                if (string.IsNullOrEmpty(runtimePackLabels))
-                {
-                    RuntimePackLabels = Array.Empty<string>();
-                }
-                else
-                {
-                    RuntimePackLabels = runtimePackLabels.Split(';');
-                }
-            }
-
-            //  The name / itemspec of the FrameworkReference used in the project
-            public string Name => _item.ItemSpec;
-
-            ////  The framework name to write to the runtimeconfig file (and the name of the folder under dotnet/shared)
-            public string LatestRuntimeFrameworkVersion => _item.GetMetadata("LatestRuntimeFrameworkVersion");
-
-            public string RuntimePackNamePatterns => _item.GetMetadata("RuntimePackNamePatterns");
-
-            public string RuntimePackRuntimeIdentifiers => _item.GetMetadata(MetadataKeys.RuntimePackRuntimeIdentifiers);
-
-            public string IsTrimmable => _item.GetMetadata(MetadataKeys.IsTrimmable);
-
-            public bool IsWindowsOnly => _item.HasMetadataValue("IsWindowsOnly", "true");
-
-            public bool RuntimePackAlwaysCopyLocal =>
-                _item.HasMetadataValue(MetadataKeys.RuntimePackAlwaysCopyLocal, "true");
-
-            public string[] RuntimePackLabels { get; }
-
-            public NuGetFramework TargetFramework { get; }
-        }
     }
 }
